Move PreviewSelectedResources to common
