+++ kolibri/plugins/coach/assets/src/views/common/resourceSelection/QuizResourceSelectionHeader.vue
@@ -21,7 +21,6 @@
         v-if="!hideSearch"
         icon="filter"
         :text="searchLabel$()"
-        @click="$emit('searchClick')"
       />
     </div>
   </div>
