+++ kolibri/plugins/coach/assets/src/views/common/resourceSelection/subPages/SelectFromBookmarks.vue
@@ -20,7 +20,7 @@
       :selectAllRules="selectAllRules"
       :selectedResources="selectedResources"
       :contentCardMessage="contentCardMessage"
-      :noSelectableResourcesIds="noSelectableResourcesIds"
+      :unselectableResourceIds="unselectableResourceIds"
       @selectResources="$emit('selectResources', $event)"
       @deselectResources="$emit('deselectResources', $event)"
       @setSelectedResources="$emit('setSelectedResources', $event)"
@@ -126,7 +126,7 @@
         type: Array,
         required: true,
       },
-      noSelectableResourcesIds: {
+      unselectableResourceIds: {
         type: Array,
         required: false,
         default: null,
